test(e2e): keep clock-skew on composed HA postgres suite, drop distribute-work .parallel
Address review feedback on the HA suite split:

- Restore the "Clock Skew and Timezone Safety" describe into the composed
  e2e_ha_full.parallel.test.ts suite, where it runs against the cluster's real
  dockerized PostgreSQL slashing-protection DB. Timezone/clock divergence between
  the node and its database can only be reproduced against a genuine, separate
  Postgres process, so the in-process PGlite variant is removed and the
  @electric-sql/pglite dependency dropped from end-to-end.

- Rename e2e_ha_distribute_work.parallel.test.ts to e2e_ha_distribute_work.test.ts:
  it has a single it(), so it needs no per-test .parallel container splitting.
