+++ yarn-project/end-to-end/src/composed/ha/e2e_ha_full.parallel.test.ts
@@ -5,17 +5,22 @@
  * and Web3Signer for remote signing. Verifies that blocks are produced,
  * attestations are signed, and no double-signing occurs.
  *
- * The cluster setup lives in `ha_full_setup.ts` and is shared with
- * `e2e_ha_distribute_work.parallel.test.ts`. The node-killing "distribute work" resilience test lives in
- * that separate file so it gets its own cluster instead of relying on running last, and the clock-skew /
- * timezone DB assertions moved to the cheap in-process `multi-node/high-availability/clock_skew.test.ts`.
+ * The cluster setup lives in `ha_full_setup.ts` and is shared with `e2e_ha_distribute_work.test.ts`. The
+ * node-killing "distribute work" resilience test lives in that separate file so it gets its own cluster
+ * instead of relying on running last. The clock-skew / timezone DB assertions stay here in the nested
+ * `Clock Skew and Timezone Safety` describe: they exercise the cluster's real dockerized PostgreSQL
+ * slashing-protection database, and timezone/clock divergence between the node and its database can only
+ * be reproduced against a genuine, separate Postgres process (not an in-process one).
  */
 import { AztecAddress, EthAddress } from '@aztec/aztec.js/addresses';
-import { SlotNumber } from '@aztec/foundation/branded-types';
+import { BlockNumber, CheckpointNumber, SlotNumber } from '@aztec/foundation/branded-types';
+import { Buffer32 } from '@aztec/foundation/buffer';
 import { retryUntil } from '@aztec/foundation/retry';
+import { sleep } from '@aztec/foundation/sleep';
 import { GovernanceProposerAbi } from '@aztec/l1-artifacts/GovernanceProposerAbi';
 import type { ValidatorClient } from '@aztec/validator-client';
-import { type DutyRow, DutyStatus } from '@aztec/validator-ha-signer/types';
+import { PostgresSlashingProtectionDatabase } from '@aztec/validator-ha-signer/db';
+import { type DutyRow, DutyStatus, DutyType } from '@aztec/validator-ha-signer/types';
 
 import { jest } from '@jest/globals';
 import { writeFile } from 'node:fs/promises';
@@ -394,5 +399,249 @@ describe('HA Full Setup', () => {
         await haNodeServices[i].reloadKeystore();
       }
     }
+  });
+
+  describe('Clock Skew and Timezone Safety', () => {
+    const rollupAddress = EthAddress.random();
+    const validatorAddress = EthAddress.random();
+    it('should not be affected by process.env.TZ changes', async () => {
+      const { mainPool } = t;
+      const spDb = new PostgresSlashingProtectionDatabase(mainPool);
+      const originalTZ = process.env.TZ;
+
+      try {
+        // Node 1 in UTC creates and signs a duty
+        process.env.TZ = 'UTC';
+        const duty1 = await spDb.tryInsertOrGetExisting({
+          rollupAddress,
+          validatorAddress,
+          slot: SlotNumber(100),
+          blockNumber: BlockNumber(0),
+          checkpointNumber: CheckpointNumber(0),
+          dutyType: DutyType.ATTESTATION,
+          messageHash: Buffer32.random().toString(),
+          nodeId: 'node-utc',
+        });
+        expect(duty1.isNew).toBe(true);
+        await spDb.updateDutySigned(
+          rollupAddress,
+          validatorAddress,
+          SlotNumber(100),
+          DutyType.ATTESTATION,
+          '0xsig',
+          duty1.record.lockToken,
+          -1,
+        );
+
+        // Wait for real database time to pass (duties need different timestamps in PostgreSQL)
+        await sleep(100);
+
+        // Node 2 in Tokyo creates and signs a duty at approximately the same time
+        process.env.TZ = 'Asia/Tokyo';
+        const duty2 = await spDb.tryInsertOrGetExisting({
+          rollupAddress,
+          validatorAddress,
+          slot: SlotNumber(101),
+          blockNumber: BlockNumber(0),
+          checkpointNumber: CheckpointNumber(0),
+          dutyType: DutyType.ATTESTATION,
+          messageHash: Buffer32.random().toString(),
+          nodeId: 'node-tokyo',
+        });
+        expect(duty2.isNew).toBe(true);
+        await spDb.updateDutySigned(
+          rollupAddress,
+          validatorAddress,
+          SlotNumber(101),
+          DutyType.ATTESTATION,
+          '0xsig',
+          duty2.record.lockToken,
+          -1,
+        );
+
+        // Verify both duties were stored at correct absolute times (seconds apart, not hours)
+        const result = await mainPool.query<{ slot: string; unix_timestamp: string }>(
+          `SELECT slot, EXTRACT(EPOCH FROM started_at) as unix_timestamp
+           FROM validator_duties
+           WHERE slot IN ('100', '101')
+           ORDER BY slot DESC`,
+        );
+
+        const timestamp1 = parseFloat(result.rows[0].unix_timestamp);
+        const timestamp2 = parseFloat(result.rows[1].unix_timestamp);
+        const diffSeconds = Math.abs(timestamp1 - timestamp2);
+
+        // Should be less than 10 seconds apart (not hours due to timezone interpretation)
+        expect(diffSeconds).toBeLessThan(10);
+      } finally {
+        process.env.TZ = originalTZ;
+      }
+    });
+
+    it('should not delete recent duties via cleanupOldDuties when node clock is ahead', async () => {
+      const { mainPool, dateProvider } = t;
+      const spDb = new PostgresSlashingProtectionDatabase(mainPool);
+
+      // Ensure clean slate for this test
+      await mainPool.query('DELETE FROM validator_duties WHERE slot = $1', ['200']);
+
+      // Create and sign a duty using our actual methods
+      const duty = await spDb.tryInsertOrGetExisting({
+        rollupAddress,
+        validatorAddress,
+        slot: SlotNumber(200),
+        blockNumber: BlockNumber(0),
+        checkpointNumber: CheckpointNumber(0),
+        dutyType: DutyType.ATTESTATION,
+        messageHash: Buffer32.random().toString(),
+        nodeId: 'test-node',
+      });
+      expect(duty.isNew).toBe(true);
+
+      await spDb.updateDutySigned(
+        rollupAddress,
+        validatorAddress,
+        SlotNumber(200),
+        DutyType.ATTESTATION,
+        '0xsig',
+        duty.record.lockToken,
+        -1,
+      );
+
+      // Verify duty exists before cleanup
+      const beforeCleanup = await mainPool.query<DutyRow>(
+        `SELECT * FROM validator_duties WHERE slot = $1 AND validator_address = $2`,
+        ['200', validatorAddress.toString().toLowerCase()],
+      );
+      expect(beforeCleanup.rows.length).toBe(1);
+      expect(beforeCleanup.rows[0].status).toBe('signed');
+
+      // Simulate node with clock 2 hours ahead using dateProvider
+      // NOTE: Database cleanup uses PostgreSQL's CURRENT_TIMESTAMP, not application time
+      // This test verifies that even if the application clock is skewed, cleanup
+      // correctly uses database time to determine duty age
+      dateProvider.setTime(Date.now() + 2 * 60 * 60 * 1000); // 2 hours ahead
+
+      try {
+        // Use our actual cleanupOldDuties method
+        const numCleaned = await spDb.cleanupOldDuties(60 * 60 * 1000); // 1 hour
+
+        // Should NOT delete the duty we just created (it uses DB's clock, not node's)
+        expect(numCleaned).toBe(0);
+      } finally {
+        // Reset dateProvider back to real time
+        dateProvider.reset();
+      }
+
+      // Verify duty still exists
+      const result = await mainPool.query<DutyRow>(
+        `SELECT * FROM validator_duties WHERE slot = $1 AND validator_address = $2`,
+        ['200', validatorAddress.toString().toLowerCase()],
+      );
+      expect(result.rows.length).toBe(1);
+    });
+
+    it('should delete old duties via cleanupOldDuties based on DB time, not node time', async () => {
+      const { mainPool, dateProvider } = t;
+      const spDb = new PostgresSlashingProtectionDatabase(mainPool);
+
+      // Ensure clean slate for this test
+      await mainPool.query('DELETE FROM validator_duties WHERE slot = $1', ['300']);
+
+      // Create and sign a duty using our actual methods
+      const duty = await spDb.tryInsertOrGetExisting({
+        rollupAddress,
+        validatorAddress,
+        slot: SlotNumber(300),
+        blockNumber: BlockNumber(0),
+        checkpointNumber: CheckpointNumber(0),
+        dutyType: DutyType.ATTESTATION,
+        messageHash: Buffer32.random().toString(),
+        nodeId: 'test-node',
+      });
+      expect(duty.isNew).toBe(true);
+
+      await spDb.updateDutySigned(
+        rollupAddress,
+        validatorAddress,
+        SlotNumber(300),
+        DutyType.ATTESTATION,
+        '0xsig',
+        duty.record.lockToken,
+        -1,
+      );
+
+      // Manually backdate the duty to 2 hours old (simulating an old duty from DB's perspective)
+      const updateResult = await mainPool.query(
+        `UPDATE validator_duties
+         SET started_at = CURRENT_TIMESTAMP - INTERVAL '2 hours',
+             completed_at = CURRENT_TIMESTAMP - INTERVAL '2 hours'
+         WHERE slot = $1 AND validator_address = $2`,
+        ['300', validatorAddress.toString().toLowerCase()],
+      );
+      expect(updateResult.rowCount).toBe(1);
+
+      // Verify duty is backdated (should be ~2 hours old)
+      const beforeCleanup = await mainPool.query<DutyRow & { age_seconds: string }>(
+        `SELECT *, EXTRACT(EPOCH FROM (CURRENT_TIMESTAMP - started_at)) as age_seconds
+         FROM validator_duties WHERE slot = $1`,
+        ['300'],
+      );
+      expect(beforeCleanup.rows.length).toBe(1);
+      expect(beforeCleanup.rows[0].status).toBe('signed');
+      expect(parseFloat(beforeCleanup.rows[0].age_seconds)).toBeGreaterThan(7000); // ~2 hours in seconds
+
+      // Simulate node with clock 1 hour behind using
+      dateProvider.setTime(Date.now() - 1 * 60 * 60 * 1000); // 1 hour behind
+
+      try {
+        // Use our actual cleanupOldDuties method - should delete based on DB time
+        const numCleaned = await spDb.cleanupOldDuties(60 * 60 * 1000); // 1 hour
+        expect(numCleaned).toBeGreaterThanOrEqual(1);
+      } finally {
+        // Reset dateProvider back to real time
+        dateProvider.reset();
+      }
+
+      // Verify duty was deleted
+      const result = await mainPool.query<DutyRow>(
+        `SELECT * FROM validator_duties WHERE slot = $1 AND validator_address = $2`,
+        ['300', validatorAddress.toString().toLowerCase()],
+      );
+      expect(result.rows.length).toBe(0);
+    });
+
+    it('should not delete recent stuck duties via cleanupOwnStuckDuties when node clock is ahead', async () => {
+      const { mainPool, dateProvider } = t;
+      const spDb = new PostgresSlashingProtectionDatabase(mainPool);
+
+      // Create a signing duty (stuck, not completed) using our actual method
+      const duty = await spDb.tryInsertOrGetExisting({
+        rollupAddress,
+        validatorAddress,
+        slot: SlotNumber(400),
+        blockNumber: BlockNumber(0),
+        checkpointNumber: CheckpointNumber(0),
+        dutyType: DutyType.ATTESTATION,
+        messageHash: Buffer32.random().toString(),
+        nodeId: 'stuck-node',
+      });
+      expect(duty.isNew).toBe(true);
+      // Don't call updateDutySigned - leave it in 'signing' state (stuck)
+
+      // Simulate node with clock 3 hours ahead
+      dateProvider.setTime(Date.now() + 3 * 60 * 60 * 1000); // 3 hours ahead
+
+      try {
+        // Use our actual cleanupOwnStuckDuties method
+        const numCleaned = await spDb.cleanupOwnStuckDuties('stuck-node', 60 * 60 * 1000); // 1 hour
+
+        // Should NOT delete the duty (it uses DB's clock, not node's)
+        expect(numCleaned).toBe(0);
+      } finally {
+        // Reset dateProvider back to real time
+        dateProvider.reset();
+      }
+    });
   });
 });
