+++ yarn-project/end-to-end/src/composed/ha/ha_full_setup.ts
@@ -2,7 +2,7 @@
  * Shared setup for the docker-compose HA full suite.
  *
  * Stands up the complete HA cluster used by `e2e_ha_full.parallel.test.ts` and
- * `e2e_ha_distribute_work.parallel.test.ts`: a bootstrap RPC/P2P node plus NODE_COUNT in-proc
+ * `e2e_ha_distribute_work.test.ts`: a bootstrap RPC/P2P node plus NODE_COUNT in-proc
  * `AztecNodeService` HA peers that share one PostgreSQL slashing-protection DB and a Web3Signer keystore.
  * Requires the docker-compose HA suite (run_test.sh ha): live Postgres (DATABASE_URL) and Web3Signer
  * sidecar.
